third_party/WebKit/LayoutTests/webaudio/decode-audio-data-basic-expected.txt

CONSOLE ERROR: line 21: Uncaught (in promise) NotSupportedError: invalid ArrayBuffer for audioData.
Basic tests for decodeAudioData function.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS decodeAudioData(null) invoked errorCallback correctly: NotSupportedError: invalid ArrayBuffer for audioData.
PASS The resources/media/24bit-44khz.wav test: successCallback has been called correctly.
PASS The resources/media/invalid-audio-file.txt test: errorCallback has been called correctly.
PASS successfullyParsed is true

TEST COMPLETE
